refactor(wasm): extract pipeline and formalize JS boundary contracts
